refactor: extension function
