+++ android/src/main/kotlin/com/ayeee/blue_print_pos/extension/Bitmap.kt
@@ -1,0 +1,11 @@
+package com.ayeee.blue_print_pos.extension
+
+import android.graphics.Bitmap
+import java.io.ByteArrayOutputStream
+
+fun Bitmap.toByteArray(): ByteArray {
+    ByteArrayOutputStream().apply {
+        compress(Bitmap.CompressFormat.PNG, 100, this)
+        return toByteArray()
+    }
+}
